RBRP
Scrip to automate extraction of entrenchment ratio tables for the state
of North Carolina (100 counties). The tables will be stored in the
master FRIS folder. The only input requirement is the path to FRIS
folder with 100 county folders each having the zip files and HEC-RAS
models in them. The tables named (CountyName_ET_Table_CSV) will be saved
within the county folders and a copy will be saved within a
sub-directory in the FRIS main folder